refactor: move ptrace engine to ptrace/engine.rs
